remove coco
Moved parse_single_instance from coco.py to api.py.
